requirements.txt
python-telegram-bot
anthropic
python-dotenv